requirements.txt
pytest == 7.3.1
selene==2.0.0b5
allure-pytest == 2.13.2
python-dotenv == 1.0.0